pyproject.toml
[project]
name = "google-adk-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.20.0",
]
